data_process/requirements.txt
acres==0.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
bitarray==3.3.1
Brotli==1.1.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
ci-info==0.3.0
ckzg==2.1.1
click==8.1.8
cloudpickle==3.1.1
colorcet==3.1.0
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cryptography==46.0.2
cssselect2==0.8.0
cycler==0.12.1
cytoolz==1.0.1
dask==2025.3.0
datashader==0.18.1
etelemetry==0.3.1
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.3.0
filelock==3.20.0
fitz==0.0.1.dev2
fonttools==4.56.0
frontend==0.0.3
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.2.4
h11==0.16.0
hexbytes==1.3.0
httplib2==0.31.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
kiwisolver==1.4.8
llvmlite==0.44.0
locket==1.0.0
looseversion==1.3.0
lxml==6.0.2
matplotlib==3.10.1
multidict==6.2.0
multipledispatch==1.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numba==0.61.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
param==2.2.1
parsimonious==0.10.0
partd==1.4.2
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.1.0
playwright==1.55.0
propcache==0.3.1
prov==2.1.1
puremagic==1.30
pyarrow==19.0.1
pycparser==2.23
pycryptodome==3.22.0
pyct==0.5.0
pydantic==2.11.3
pydantic_core==2.33.1
pydot==4.0.1
pydyf==0.11.0
pyee==13.0.0
pyparsing==3.2.1
pypdfium2==4.30.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.1
pyunormalize==16.0.0
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.2.1
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
scipy==1.15.3
seaborn==0.13.2
simplejson==3.20.2
six==1.16.0
sniffio==1.3.1
soupsieve @ file:///croot/soupsieve_1696347547217/work
starlette==0.48.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toolz==1.0.0
traits==7.0.2
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.37.0
weasyprint==66.0
web3==7.10.0
webencodings==0.5.1
websockets==15.0.1
xarray==2025.6.1
yarl==1.19.0
zipp==3.21.0
zopfli==0.2.3.post1